+++ src/ng-select/ng-select.component.spec.ts
@@ -1557,6 +1557,32 @@ describe('NgSelectComponent', function () {
             expect(selectClasses.contains('ng-select-top')).toBeFalsy();
             expect(panelClasses.contains('ng-select-top')).toBeFalsy();
         }));
+
+        it('should return current panel position', fakeAsync(() => {
+            const fixture = createTestingModule(
+                NgSelectTestCmp,
+                `<ng-select [items]="cities" appendTo="body"></ng-select>`);
+
+            const select = fixture.componentInstance.select;
+            select.open();
+            tickAndDetectChanges(fixture);
+
+            expect(select.currentPanelPosition).toBe('bottom');
+        }));
+
+        it('should return undefined for current panel position if dropdown is closed', fakeAsync(() => {
+            const fixture = createTestingModule(
+                NgSelectTestCmp,
+                `<ng-select [items]="cities" appendTo="body"></ng-select>`);
+
+            const select = fixture.componentInstance.select;
+            select.open();
+            tickAndDetectChanges(fixture);
+            select.close();
+            tickAndDetectChanges(fixture);
+
+            expect(select.currentPanelPosition).toBeUndefined();
+        }));
     });
 
     describe('Custom templates', () => {
@@ -1854,10 +1880,7 @@ describe('NgSelectComponent', function () {
             }));
 
             it('should not open dropdown when maximum of items is reached', fakeAsync(() => {
-                const clickArrow = () => arrowIcon.triggerEventHandler('click', {
-                    stopPropagation: () => {
-                    }
-                });
+                const clickArrow = () => arrowIcon.triggerEventHandler('click', {});
                 selectOption(fixture, KeyCode.ArrowDown, 0);
                 selectOption(fixture, KeyCode.ArrowDown, 1);
                 tickAndDetectChanges(fixture);
@@ -3365,8 +3388,6 @@ function createTestingModule<T>(cmp: Type<T>, template: string): ComponentFixtur
 
 function createEvent(target = {}) {
     return {
-        stopPropagation: () => {
-        },
         preventDefault: () => {
         },
         target: {